myproject/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
chart-studio==1.1.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
crispy-bootstrap4==2023.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-braces==1.15.0
django-crispy-forms==2.1
django-gravatar2==1.4.4
django-model-utils==4.3.1
django-star-ratings==0.9.2
django-widget-tweaks==1.5.0
executing==2.0.1
fonttools==4.44.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
install==1.3.5
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
retrying==1.3.4
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
swapper==1.3.0
swig==4.1.1.post0
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.9
wheel==0.41.3
